split PathJob into model and gui; Made PathJob a regular feature and moved operations into child compound; Added stock feature.
