+++ src/Mod/Path/PathScripts/PathDressupRampEntry.py
@@ -563,7 +563,7 @@ class ViewProviderDressup:
         '''this makes sure that the base operation is added back to the project and visible'''
         FreeCADGui.ActiveDocument.getObject(arg1.Object.Base.Name).Visibility = True
         job = PathUtils.findParentJob(self.obj)
-        PathUtils.addObjectToJob(arg1.Object.Base, job)
+        job.addOperation(arg1.Object.Base)
         arg1.Object.Base = None
         return True
